Move getSummonedCreature() to summoning
